dom/tests/mochitest/devicestorage/test_overwrite.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=717103
-->
<head>
  <title>Test for basic sanity of the device storage API </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="devicestorage_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=717103">Mozilla Bug 717103</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

var filename = "devicestorage/aaaa"

devicestorage_setup();


function deleteSuccess(e) {
  devicestorage_cleanup();
}

function deleteError(e) {
  ok(false, "deleteError was called : " + e.target.error.name);
  devicestorage_cleanup();
}

function addOverwritingSuccess(e) {
  ok(false, "addOverwritingSuccess was called.");
}

function addOverwritingError(e) {
  ok(true, "Adding to the same location should fail");

  var storage = navigator.getDeviceStorage("profile");
  request = storage[0].delete(filename)
  request.onsuccess = deleteSuccess;
  request.onerror = deleteError;
}

function addSuccess(e) {
  ok(true, "addSuccess was called");

  var storage = navigator.getDeviceStorage("profile");
  ok(navigator.getDeviceStorage, "Should have getDeviceStorage");

  request = storage[0].addNamed(createRandomBlob(), filename);
  ok(request, "Should have a non-null request");

  request.onsuccess = addOverwritingSuccess;
  request.onerror = addOverwritingError;
}

function addError(e) {
  // test file is already exists.  clean it up and try again..
  var storage = navigator.getDeviceStorage("profile");
  request = storage[0].delete(filename)
  request.onsuccess = runtest;
}

function runtest() {
  var storage = navigator.getDeviceStorage("profile");
  ok(navigator.getDeviceStorage, "Should have getDeviceStorage");

  request = storage[0].addNamed(createRandomBlob(), filename);
  ok(request, "Should have a non-null request");

  request.onsuccess = addSuccess;
  request.onerror = addError;
}

runtest();

</script>
</pre>
</body>
</html>
